Moved existing org logic in connector-thegraph to contained folder
