+++ packages/connector-thegraph/src/core/GraphQLWrapper.ts
@@ -1,3 +1,4 @@
+import 'isomorphic-unfetch';
 import { Client } from '@urql/core'
 import { DocumentNode } from 'graphql';
 import {
