+++ packages/connector-thegraph/src/org/graphql/fragments.ts
