+++ packages/connector-thegraph/src/org/graphql/queries.ts
