+++ packages/connector-thegraph/src/org/graphql/types.ts
