+++ packages/connector-thegraph/src/org/parse/roles.ts
