+++ packages/connector-thegraph/src/types.ts
@@ -1,5 +1,3 @@
-import ConnectorTheGraphBase from "./core/ConnectorTheGraphBase"
-
 export type QueryResult = any
 export type DataGql = any
